+++ javascript/pages/VideoGalleryPage.js
@@ -1,0 +1,10 @@
+(function($){
+	"use strict";
+	$(document).ready(function(){
+		$("a.yt-video").each(function(){
+			$(this).fancybox({
+				caption : $(this).siblings('.vid-description').text()
+			});			
+		});
+	});
+}(jQuery));