+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-## Package: iqnection-silverstripe-3-pages-videogallerypage
+## Package: iqnection-silverstripe-pages-videogallerypage
 # Change Log
 
+
+## [2.0.0]
+- updated for SilverStripe 4
+- now incorporates YouTube's API for dynamic updating
+- instead of entering embed code, you can just enter the YouTube share link
+- uses Fancybox 3 for better experience and faster page loading (no on page iframes)
 
 ## [1.0.1]
 - Update for compatibility with SS 3.3+